✨ feat(project-detail): enhance project detail component layout and functionality by adding detailed project information, improving contact persons display, and integrating contract staff service for better user experience
