+++ research-indicators/src/app/pages/platform/pages/project-detail/components/project-dashboard/project-dashboard.component.html
@@ -13,106 +13,6 @@
   <div class="grid grid-cols-1 gap-5 lg:grid-cols-[3fr_1fr]  lg:min-h-[520px] lg:items-stretch ">
     <div class="flex min-w-0 flex-col gap-5 lg:h-full lg:self-stretch ">
 
-      <!-- Project information -->
-      <section class="rounded-xl border border-[#e8ebed] bg-white p-5 shadow-xs">
-        <div class="mb-4 flex flex-col gap-3 sm:flex-row sm:items-start sm:justify-between">
-          <div class="flex flex-col gap-1 pb-1">
-            <span class="font-['Barlow'] text-[15px] font-[550] leading-tight text-[#112f5c]">Project information</span>
-            <span class="font-['Barlow'] text-[14px] leading-4 text-[#777C83]">Project dates, budget, donor, unit, and
-              contact persons.</span>
-          </div>
-          <div class="flex shrink-0 flex-col sm:items-end">
-            <span
-              class="font-['Space_Grotesk'] text-[11px] font-[600] uppercase tracking-wide text-[#8D9299]">Budget</span>
-            <strong class="font-['Barlow'] text-[20px] font-semibold leading-none tabular-nums text-[#112F5C]">{{
-              projectGrantAmount() }} USD</strong>
-          </div>
-        </div>
-
-        <div class="grid grid-cols-1 gap-3 md:grid-cols-4">
-          <div class="rounded-lg border border-[#e8ebed] bg-white px-4 py-1.5 flex flex-col justify-center">
-            <span
-              class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Start
-              date</span>
-            <strong class="block font-['Barlow'] text-[14px] font-semibold text-[#112F5C]">{{ (project().start_date |
-              date: 'dd/MM/yyyy') || 'Not defined' }}</strong>
-          </div>
-
-          <div class="rounded-lg border border-[#e8ebed] bg-white px-4 py-1.5 flex flex-col justify-center">
-            <span class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">End
-              date</span>
-            <strong class="block font-['Barlow'] text-[14px] font-semibold text-[#112F5C]">{{ (project().end_date |
-              date: 'dd/MM/yyyy') || 'Not defined' }}</strong>
-          </div>
-
-          <div class="rounded-lg border border-[#e8ebed] bg-white px-4 py-1.5 flex flex-col justify-center">
-            <span
-              class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Extension
-              date</span>
-            <strong class="block font-['Barlow'] text-[14px] font-semibold text-[#112F5C]">{{
-              (project().extension_date | date: 'dd/MM/yyyy') || 'Not defined' }}</strong>
-          </div>
-
-          <div class="rounded-lg border border-[#e8ebed] bg-white px-4 py-1.5 flex flex-col justify-center">
-            <span
-              class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Lever</span>
-            <strong class="block truncate font-['Barlow'] text-[16.5px] font-semibold text-[#345B8F]"
-              [title]="projectLeverName()">{{ projectLeverName() }}</strong>
-          </div>
-        </div>
-
-        <div class="mt-3 grid grid-cols-1 gap-3 md:grid-cols-3">
-          <div class="rounded-lg border border-[#e8ebed] bg-white px-4 py-1.5 flex flex-col justify-center">
-            <span
-              class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Donor</span>
-            <strong class="block truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]"
-              [title]="project().donor || '—'">{{ project().donor || '—' }}</strong>
-          </div>
-
-          <div class="rounded-lg border border-[#e8ebed] bg-white px-4 py-1.5 flex flex-col justify-center">
-            <span
-              class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Division</span>
-            <strong class="block truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]"
-              [title]="projectDivisionLabel()">{{ projectDivisionLabel() }}</strong>
-          </div>
-
-          <div class="rounded-lg border border-[#e8ebed] bg-white px-4 py-1.5 flex flex-col justify-center">
-            <span
-              class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Unit</span>
-            <strong class="block truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]"
-              [title]="projectUnitLabel()">{{ projectUnitLabel() }}</strong>
-          </div>
-        </div>
-
-        <div class="mt-5 border-t border-[#e8ebed] pt-2">
-          <span
-            class="!font-['Space_Grotesk'] text-[11px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Contact
-            persons</span>
-
-          @if (staffEmpty()) {
-          <p class="m-0 mt-3 font-['Barlow'] text-[13px] text-[#777C83]">No contact persons were found for this project.
-          </p>
-          } @else {
-          <div class="mt-3 grid grid-cols-1 gap-3 md:grid-cols-2 xl:grid-cols-3">
-            @for (person of contractStaff.staff(); track person.name + person.role) {
-            <article class="flex min-w-0 items-center gap-3 rounded-lg border border-[#e8ebed] bg-white p-4">
-              <div
-                class="flex h-10 w-10 shrink-0 items-center justify-center rounded-full bg-[#E8F0F7] font-['Barlow'] text-[13px] font-semibold uppercase text-[#345B8F]">
-                {{ getContactInitials(person.name) }}
-              </div>
-              <div class="min-w-0 flex flex-col gap-0.5">
-                <span class="truncate font-['Barlow'] text-[14px] font-semibold leading-snug text-[#112F5C]"
-                  [title]="person.name">{{ person.name }}</span>
-                <span class="truncate font-['Barlow'] text-[13px] leading-snug text-[#777C83]" [title]="person.role">{{
-                  person.role }}</span>
-              </div>
-            </article>
-            }
-          </div>
-          }
-        </div>
-      </section>
-
       <!-- Results by indicator -->
       <div class="grid grid-cols-1 gap-5 lg:grid-cols-2 lg:items-stretch">
 
@@ -156,6 +56,8 @@
           </app-project-dashboard-card>
         </div>
       </div>
+
+      <app-geo-scope-card class="h-full w-full"></app-geo-scope-card>
     </div>
     <div class="flex h-full min-w-0 flex-col gap-5 lg:min-h-0 lg:self-stretch">
       <section class="flex shrink-0 flex-col gap-4 rounded-xl border border-[#e8ebed] bg-white p-5 shadow-xs">
@@ -264,8 +166,6 @@
     </div>
   </div>
   <!-- Pending revision results -->
-  <app-geo-scope-card class="h-full w-full"></app-geo-scope-card>
-
   <section
     class="flex min-h-0 flex-col overflow-hidden rounded-xl border border-[#e8ebed] bg-white shadow-xs lg:flex-1">
     <header class="border-b border-[#e8ebed] px-5 py-4">
