+++ research-indicators/src/app/pages/platform/pages/project-detail/project-detail.component.html
@@ -1,15 +1,116 @@
 <div class="flex h-full min-w-0 flex-col">
   <div class="w-full min-w-0 pt-[8px] sm:pt-[12px] bg-white md:pt-[20px]">
-    <div class="mb-4 px-[24px] md:px-[28px] py-[9px]">
-      <span class="m-0 mt-2 font-['Barlow'] text-[18px] font-semibold leading-[20px] text-[#112F5C]">
-        {{ currentProject().agreement_id }} - {{ currentProject().description }}
-      </span>
-      <div class="flex flex-col gap-1 pt-1">
-        <span class="font-['Barlow'] text-[15px] font-normal leading-[18px] text-[#8D9299]">
-          {{ currentProject().department }} Department
-        </span>
-        <app-custom-tag class="shrink-0" [statusId]="projectStatus().statusId"
-          [statusName]="projectStatus().statusName"></app-custom-tag>
+    <div class="mb-4 px-[24px] py-[9px] md:px-[28px]">
+      <div class="flex flex-col gap-7 lg:flex-row lg:items-stretch lg:gap-10">
+        <header class="min-w-0 flex-1 border-b border-[#E8EBED] pb-4 lg:basis-[80%] lg:border-b-0 lg:border-r lg:pb-0 lg:pr-10">
+          <div class="flex flex-col gap-3">
+            <div class="min-w-0 flex flex-col gap-1">
+              <span class="m-0 font-['Barlow'] text-[18px] font-semibold leading-[20px] text-[#112F5C]">
+                {{ currentProject().agreement_id }} - {{ currentProject().description }}
+              </span>
+              <span class="font-['Barlow'] text-[15px] font-normal leading-[18px] text-[#8D9299]">
+                {{ currentProject().department }} Department
+              </span>
+              <app-custom-tag class="shrink-0" [statusId]="projectStatus().statusId"
+                [statusName]="projectStatus().statusName"></app-custom-tag>
+            </div>
+          </div>
+
+          <dl class="mt-4 grid grid-cols-1 gap-x-6 gap-y-2 md:grid-cols-2 xl:grid-cols-5">
+            <div class="min-w-0">
+              <dt
+                class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Budget</dt>
+              <dd class="m-0 truncate font-['Barlow'] text-[14px] font-semibold tabular-nums text-[#112F5C]">{{
+                projectGrantAmount() }} USD</dd>
+            </div>
+
+            <div class="min-w-0">
+              <dt
+                class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Start
+                date</dt>
+              <dd class="m-0 truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]">{{ (currentProject().start_date
+                | date: 'dd/MM/yyyy') || 'Not defined' }}</dd>
+            </div>
+
+            <div class="min-w-0">
+              <dt class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">End
+                date</dt>
+              <dd class="m-0 truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]">{{ (currentProject().end_date |
+                date: 'dd/MM/yyyy') || 'Not defined' }}</dd>
+            </div>
+
+            <div class="min-w-0">
+              <dt
+                class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Extension
+                date</dt>
+              <dd class="m-0 truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]">{{
+                (currentProject().extension_date | date: 'dd/MM/yyyy') || 'Not defined' }}</dd>
+            </div>
+
+            <div class="min-w-0">
+              <dt
+                class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Lever</dt>
+              <dd class="m-0 truncate font-['Barlow'] text-[14px] font-semibold text-[#345B8F]"
+                [title]="projectLeverName()">{{ projectLeverName() }}</dd>
+            </div>
+          </dl>
+
+          <dl class="mt-3 grid grid-cols-1 gap-x-6 gap-y-2 md:grid-cols-3">
+            <div class="min-w-0">
+              <dt
+                class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Donor</dt>
+              <dd class="m-0 truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]"
+                [title]="currentProject().donor || '—'">{{ currentProject().donor || '—' }}</dd>
+            </div>
+
+            <div class="min-w-0">
+              <dt
+                class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Division</dt>
+              <dd class="m-0 truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]"
+                [title]="projectDivisionLabel()">{{ projectDivisionLabel() }}</dd>
+            </div>
+
+            <div class="min-w-0">
+              <dt
+                class="!font-['Space_Grotesk'] text-[10px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Unit</dt>
+              <dd class="m-0 truncate font-['Barlow'] text-[14px] font-semibold text-[#112F5C]"
+                [title]="projectUnitLabel()">{{ projectUnitLabel() }}</dd>
+            </div>
+          </dl>
+        </header>
+
+        <aside class="min-w-0 lg:basis-[20%] lg:pl-4">
+          <div class="mb-3 flex items-center justify-between gap-3">
+            <span
+              class="!font-['Space_Grotesk'] text-[11px] font-[600] uppercase tracking-[0.03rem] text-[#8D9299]">Contact
+              persons</span>
+            @if (contractStaff.loading()) {
+            <span class="font-['Barlow'] text-[12px] text-[#777C83]">Loading...</span>
+            }
+          </div>
+
+          @if (staffEmpty()) {
+          <p class="m-0 font-['Barlow'] text-[13px] leading-snug text-[#777C83]">No contact persons were found for this
+            project.</p>
+          } @else {
+          <div class="flex max-h-[132px] flex-col gap-2 overflow-y-auto pr-1">
+            @for (person of contractStaff.staff(); track person.name + person.role) {
+            <article class="flex min-w-0 items-center gap-3">
+              <div
+                class="flex h-9 w-9 shrink-0 items-center justify-center rounded-full bg-[#E8F0F7] font-['Barlow'] text-[12px] font-semibold uppercase text-[#345B8F]">
+                {{ getContactInitials(person.name) }}
+              </div>
+              <div class="min-w-0 flex flex-col gap-0.5">
+                <span class="truncate font-['Barlow'] text-[13px] font-semibold leading-snug text-[#112F5C]"
+                  [title]="person.name">{{ person.name }}</span>
+                <span class="truncate font-['Barlow'] text-[12px] leading-snug text-[#777C83]" [title]="person.role">{{
+                  person.role }}</span>
+              </div>
+            </article>
+            }
+          </div>
+          }
+        </aside>
       </div>
     </div>
